Add files via upload
The BookList App is simply in HTML, CSS, and JavaScript. Taking about the features of this system, the user can make the list of books with their authors and keep them as records. You just have to type the book information in the text fields and click on the submit button to add the information on the record. This project includes a lot of javascript for making the functioning of the project.